Replace ProcessorWorkingBeatmap with version in base project
